+++ cvat/apps/auto_annotation/apps.py
@@ -1,0 +1,11 @@
+
+# Copyright (C) 2018 Intel Corporation
+#
+# SPDX-License-Identifier: MIT
+
+from django.apps import AppConfig
+
+
+class AutoAnnotationConfig(AppConfig):
+    name = "auto_annotation"
+