+++ cvat/urls.py
@@ -37,6 +37,9 @@ urlpatterns = [
 if apps.is_installed('cvat.apps.tf_annotation'):
     urlpatterns.append(path('tensorflow/annotation/', include('cvat.apps.tf_annotation.urls')))
 
+if apps.is_installed('cvat.apps.auto_annotation'):
+    urlpatterns.append(path('auto_annotation/', include('cvat.apps.auto_annotation.urls')))
+
 if apps.is_installed('cvat.apps.log_viewer'):
     urlpatterns.append(path('analytics/', include('cvat.apps.log_viewer.urls')))